rename run_matbench_task() to run_wrenformer() and refactor to support arbitrary datasets
remove flake8 max complexity = 20 from setup.cfg due to aviary/wrenformer/run.py:45:1: C901 'run_wrenformer' is too complex (24)
