+++ setup.cfg
@@ -1,6 +1,5 @@
 [flake8]
 max-line-length = 100
-max-complexity = 20
 # E203: whitespace before ':' preferred by black
 ignore = E203
